Inline test network creation in some tests
